requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# UI Framework
streamlit>=1.28.0

# Image Processing Utilities
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0

# Training Utilities
tqdm>=4.66.0

# Development Tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
